obtain product from db

+++ site/src/jsMain/kotlin/org/fedsal/buenpuerto/pages/Index.kt
@@ -1,11 +1,15 @@
 package org.fedsal.buenpuerto.pages
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import com.stevdza.san.kotlinbs.components.BSSpinner
+import com.stevdza.san.kotlinbs.models.SpinnerStyle
+import com.stevdza.san.kotlinbs.models.SpinnerVariant
 import com.varabyte.kobweb.compose.foundation.layout.Arrangement
 import com.varabyte.kobweb.compose.foundation.layout.Box
 import com.varabyte.kobweb.compose.foundation.layout.Column
@@ -14,21 +18,22 @@ import com.varabyte.kobweb.compose.ui.Alignment
 import com.varabyte.kobweb.compose.ui.Modifier
 import com.varabyte.kobweb.compose.ui.modifiers.fillMaxSize
 import com.varabyte.kobweb.core.Page
-import org.fedsal.buenpuerto.data.datasource.supabase.SupabaseOrderRemoteDataSource
+import com.varabyte.kobweb.core.rememberPageContext
 import org.fedsal.buenpuerto.data.datasource.OrderLocalDataSource
+import org.fedsal.buenpuerto.data.datasource.supabase.SupabaseOrderRemoteDataSource
 import org.fedsal.buenpuerto.data.repository.OrderRepository
 import org.fedsal.buenpuerto.domain.model.Order
 import org.fedsal.buenpuerto.domain.model.OrderItem
-import org.fedsal.buenpuerto.domain.model.Product
 import org.fedsal.buenpuerto.sections.CheckoutSection
 import org.fedsal.buenpuerto.sections.FooterSection
 import org.fedsal.buenpuerto.sections.HeaderSection
 import org.fedsal.buenpuerto.sections.ImageCarouselSection
 import org.fedsal.buenpuerto.sections.ProductPriceSection
 import org.fedsal.buenpuerto.sections.ProductTitleSection
+import org.fedsal.buenpuerto.utils.Res
 import org.fedsal.buenpuerto.viewmodel.OrderViewModel
 
-@Page
+@Page("{productCode}")
 @Composable
 fun HomePage() {
     var cachedOrder by remember { mutableStateOf(Order()) }
@@ -54,45 +59,53 @@ fun HomePage() {
         )
     )
 
+    val ctx = rememberPageContext()
+    val productCode = ctx.route.params.getValue(Res.QueryParams.PRODUCT_CODE)
+
+    LaunchedEffect(productCode) {
+        viewModel.initializeData(productCode)
+    }
+
     val uiState = viewModel.uiState.collectAsState()
 
     var menuOpened by remember { mutableStateOf(false) }
-    val product = Product(
-        id = 1,
-        code = "123",
-        name = "Lorem ipsum dolor sit amet consectetur adipiscing elit sed do eiusmod tempor incididunt ut labore et dolore magna aliqua",
-        imagesUrl = listOf("https://http2.mlstatic.com/D_NQ_NP_2X_661732-MLA82942925465_032025-F.webp", "https://example.com/image2.jpg"),
-        price = 11122000.0
-    )
     Box(modifier = Modifier.fillMaxSize()) {
-        Column(
-            Modifier.fillMaxSize(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-            verticalArrangement = Arrangement.Top
-        ) {
-            HeaderSection(onBagClicked = { menuOpened = true })
-            ImageCarouselSection(product.imagesUrl)
-            ProductTitleSection(product.name)
-            ProductPriceSection(product.price)
-            Spacer()
-            FooterSection { quantity ->
-                val orderItem = OrderItem(
-                    product = product,
-                    quantity = quantity
-                )
-                viewModel.addItem(orderItem)
-            }
-        }
-        if (menuOpened) {
-            CheckoutSection(onMenuClosed = { menuOpened = false }, order = uiState.value.order,
-                onDecrement = { viewModel.removeItem(it) },
-                onIncrement = { viewModel.addItem(it) },
-                onRemove = { viewModel.deleteItem(it) },
-                onPlaceOrder = {
-
-                    viewModel.sendOrder("Federico")
-                }
+        if (uiState.value.isLoading) {
+            BSSpinner(
+                modifier = Modifier.align(Alignment.Center),
+                variant = SpinnerVariant.Large,
+                style = SpinnerStyle.Dark
             )
+        } else {
+            Column(
+                Modifier.fillMaxSize(),
+                horizontalAlignment = Alignment.CenterHorizontally,
+                verticalArrangement = Arrangement.Top
+            ) {
+                HeaderSection(onBagClicked = { menuOpened = true })
+                ImageCarouselSection(uiState.value.product.imagesUrl)
+                ProductTitleSection(uiState.value.product.name)
+                ProductPriceSection(uiState.value.product.price)
+                Spacer()
+                FooterSection { quantity ->
+                    val orderItem = OrderItem(
+                        product = uiState.value.product,
+                        quantity = quantity
+                    )
+                    viewModel.addItem(orderItem)
+                }
+            }
+            if (menuOpened) {
+                CheckoutSection(onMenuClosed = { menuOpened = false }, order = uiState.value.order,
+                    onDecrement = { viewModel.removeItem(it) },
+                    onIncrement = { viewModel.addItem(it) },
+                    onRemove = { viewModel.deleteItem(it) },
+                    onPlaceOrder = {
+
+                        viewModel.sendOrder("Federico")
+                    }
+                )
+            }
         }
     }
 }
